Move refPropagation into its own file
